Migrate to PostgreSQL

+++ Warehouse/Warehouse.Api/Warehouse.Infrastructure/Persistence/EntitiesConfiguration/CategoryTypeConfiguration.cs
@@ -29,7 +29,7 @@ namespace Warehouse.Infrastructure.Persistence.EntitiesConfiguration
 
             entityTypeBuilder
                 .Property(nameof(Category.State))
-                .HasColumnType("tinyint")
+                .HasColumnType("smallint")
                 .HasColumnName(nameof(Category.State))
                 .IsRequired();
